+++ frontend/nametables/nametable.cpp
@@ -3,6 +3,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <assert.h>
+#include <limits.h>
 
 name_table* NameTableInit(){
     name_table* nametab = (name_table*)calloc(sizeof(name_table), 1);
@@ -20,7 +21,7 @@ name_table* NameTableInit(){
     return nametab;
 }
 
-size_t NameTableAddName(name_table* nametab, char* num_of_variable){
+int NameTableAddName(name_table* nametab, char* num_of_variable){
     assert(nametab);
     assert(num_of_variable);
 
@@ -46,18 +47,18 @@ void NameTableRealloc(name_table* nametab, size_t num_of_elem){
     }
 }
 
-size_t FindVarInNameTable(name_table* nametab, char* num_of_variable){
-    for(size_t name = 0; name < nametab->first_free; name++){
+int FindVarInNameTable(name_table* nametab, char* num_of_variable){
+    for(int name = 0; name < nametab->first_free; name++){
         if(!strcmp(nametab->var_info[name].variable_name, num_of_variable)){
             return name;
         }
     }
-    return SIZE_MAX;
+    return INT_MAX;
 }
 
 void NameTableDestroy(name_table* nametab){
     if(nametab){
-        for(size_t i = 0; i < nametab->first_free; i++){
+        for(int i = 0; i < nametab->first_free; i++){
             free(nametab->var_info[i].variable_name);
         }
         if(nametab->var_info){
