+++ stack/stack.h
@@ -3,7 +3,7 @@
 #include <stdlib.h>
 #include <unistd.h> 
 
-typedef ssize_t stack_elem_t; 
+typedef int stack_elem_t; 
 
 const size_t FRONT_CANARY_STACK = 0x4255463132333445;
 const size_t TAIL_CANARY_STACK = 0x4E45534255464145;
